+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 
 #### Additions
 * Added Bee Armor (1.15+ only)
+* (API) Added Adventure text-library
 
 #### Changes
 * Performance optimizations for Cargo networks
@@ -44,6 +45,7 @@
 * Fixed #2664
 * Fixed #2655
 * Fixed #2675
+* Fixed /sf timings --verbose not working correctly
 
 ## Release Candidate 18 (03 Dec 2020)